Create breakpoints in percentage
I assigned a percentage value to each class used here, such as using w-5 to represent width: 5% (relative to its parent).
and I did the same for medium and larger screens by using the "md" class. For example, if you use "w-md-5" as a class, it means 5% width of its parent from medium and above medium.